pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "viplanner"
version = "0.1.0"
description = "Visual Imperative Planner for Legged Robots"
authors = [{name = "Pascal Roth", email = "rothpa@ethz.ch"}]
license = {file = "LICENSE.txt"}
readme = "README.md"
requires-python = ">=3.7"
keywords = ["robotics", "planning", "legged-robots"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
dependencies = [
    "torch",
    "torchvision",
    "PyYAML==6.0",
    "tqdm",
    "matplotlib",
    "networkx",
    "scipy",
    "open3d==0.17.0",
    "wandb==0.14.0",
    "opencv-python-headless",
]

[project.optional-dependencies]
inference = [
    "mmcv==2.0.0",
    "mmengine",
    "mmdet",
]
standard = [
    "pypose",
]
jetson = [
    "torch==1.11",
]

[project.urls]
homepage = "https://github.com/pascal-roth/viplanner"
repository = "https://github.com/pascal-roth/viplanner.git"

[tool.setuptools.packages]
find = {}
